[CodingStandard] decouple itemCount method to ArrayTokensAnalyzer
